+++ src/components/Chair/index.tsx
