+++ src/renderer/App.css
@@ -14,8 +14,3 @@ body {
   justify-content: center;
   align-items: center;
 }
-/*
-#scene{
-  width: 400px;
-  height: 400px;
-} */
